# api.yml
openapi: 3.0.0
info:
    title: OmniStudy API
    description: API for the OmniStudy application
    version: 1.0.0

servers:
    - url: http://localhost:3001
      description: Development, staging, and production server (app is dockerized in one VM)

paths:
    /users:
        get: 
            summary: Get all users
            responses: 
                '200':
                    description: Users retrieved successfully
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    users:
                                        type: array
                                        items:
                                            $ref: '#/components/schemas/User'
                                    message:
                                        description: Success message
                                        type: string
                '404':
                    description: No users found
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    users:
                                        type: array
                                        items: {}
                                    message:
                                        description: Error message
                                        type: string
                '500':
                    description: Internal server error (likely due to a query/connection error)
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    users:
                                        type: array
                                        items: {}
                                    message:
                                        description: Error message
                                        type: string
        post:
            summary: Create a new user
            description: Same as POST /register
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                firstName:
                                    description: First name of the user
                                    type: string
                                lastName:
                                    description: Last name of the user
                                    type: string
                                email:
                                    description: Email of the user
                                    type: string
                                username:
                                    description: Username of the user
                                    type: string
                                password:
                                    description: Password of the user
                                    type: string
                                phone:
                                    description: Phone number of the user
                                    type: string
                        examples:
                            with_phone:
                                value:
                                    firstName: Jamison
                                    lastName: Grudem
                                    email: info@omnistudy.io
                                    username: jgrudem
                                    password: password
                                    phone: 123 456 7890
                            without_phone:
                                value:
                                    first_name: Jamison
                                    last_name: Grudem
                                    email: info@omnistudy.io
                                    username: jgrudem
                                    password: password
            responses:
                '200':
                    description: User created successfully
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    user:
                                        $ref: '#/components/schemas/User'
                                    message:
                                        description: Success message
                                        type: string
    /users/{userId}:
        get:
            summary: Get a user by ID
            parameters:
                - in: path
                  name: userId
                  required: true
                  description: ID of the user to retrieve
                  schema:
                    type: integer
            responses:
                '200':
                    description: User retrieved successfully
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    user:
                                        $ref: '#/components/schemas/User'
                                    message:
                                        description: Success message
                                        type: string
                                    code:
                                        description: HTTP status code
                                        type: integer
                '404':
                    description: User not found
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    user:
                                        nullable: true
                                    message:
                                        description: Error message
                                        type: string
                                    code:
                                        description: HTTP status code
                                        type: integer
                '500':
                    description: Internal server error (likely due to a query/connection error)
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    user:
                                        nullable: true
                                    message:
                                        description: Error message
                                        type: string
                                    code:
                                        description: HTTP status code
                                        type: integer

components:
    schemas:
        User:
            properties:
                id: 
                    description: Auto-incremented unique identifier of the user
                    type: integer
                api_key:
                    description: Unique token used for authentication
                    type: string
                name: 
                    description: Full name of the user
                    type: string
                first_name: 
                    description: First name of the user
                    type: string
                last_name:
                    description: Last name of the user
                    type: string
                username:
                    description: Username of the user
                    type: string  
                email:
                    description: Email of the user
                    type: string
                password:
                    description: Hashed password of the user
                    type: string
                phone:
                    description: Phone number of the user
                    type: string
                created_at:
                    description: Date and time the user was created
                    type: string
                online:
                    description: Online status (true or false)
                    type: boolean